feat(grafana): Add script to run a node with docker and monitoring

+++ docker/grafana-local/run.sh
@@ -1,11 +1,26 @@
 #!/bin/bash
 
-cleanup() {  
-  echo "Stopping services..."
+# shut down running services
+if docker compose ps | grep -q "Up"; then
+  echo "Grafana Docker Compose already running, restarting..."
+
   docker compose down
-}
+fi
 
-trap cleanup EXIT
+if [[ "$1" == "--cluster" ]]; then
+  PROMETHEUS_CONFIG_FILE="prometheus-cluster.yml"
+  echo "Running for cluster in a detached mode"
+  docker compose up -d
+# adds monitoring network to work with node started with docker
+elif [[ "$1" == "--network-override" ]]; then
+  echo "Running grafana for fullnode in a detached mode"
+  PROMETHEUS_CONFIG_FILE="prometheus.yaml"
+  docker compose -f docker-compose.yaml -f docker-compose-grafana-override.yaml up -d
+else
+  echo "Running in a detached mode"
+  PROMETHEUS_CONFIG_FILE="prometheus.yaml"
+  docker compose up -d
+fi
 
-# Run docker-compose with the updated configuration file
-docker compose up
+echo "Use 'docker compose down' to stop services"
+
